buysell/requirements.txt
Django
mysqlclient
coverage
djangorestframework
django-cors-headers